requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2
django-cors-headers==4.2.0
psycopg2-binary==2.9.6
celery==5.3.1
redis==4.5.5
channels==4.0.0
channels-redis==4.1.0
django-redis==5.4.0